+++ src/test/java/com/fullstackmarc/fwkpoc/selenium/pages/google/MyFiddlePage.java
@@ -1,4 +1,4 @@
-package com.fullstackmarc.fwkpoc.selenium.pages.myfiddle;
+package com.fullstackmarc.fwkpoc.selenium.pages.google;
 
 import com.fullstackmarc.fwkpoc.selenium.pages.InputData;
 import com.fullstackmarc.fwkpoc.selenium.pages.Named;
